+++ conf/conf.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-const Version = "7.14.0"
+const Version = "7.15.0"
 
 const (
 	CONTENT_TYPE_JSON      = "application/json"